Change working dir on docker image
The expected command to use for that image is

```
docker run --rm -it -v $(pwd):/data parsec-docs
```

Where your current working directory is `parsec-cloud` root.

After that, docker open a interactive shell whose working dir is `/data` but to actually execute the `make` to build the documentation,
you would have to cd to `/data/docs`.

With this change, you no longer need to do the `cd` part.